[CHORE] 왼쪽 사이드 바 책임 멤버로 변경
