+++ src/main/java/com/tiki/server/member/controller/docs/MemberControllerDocs.java
@@ -4,12 +4,17 @@ import com.tiki.server.common.dto.BaseResponse;
 import com.tiki.server.common.dto.ErrorResponse;
 import com.tiki.server.common.dto.SuccessResponse;
 import com.tiki.server.member.dto.request.MemberProfileCreateRequest;
+import com.tiki.server.memberteammanager.dto.response.BelongTeamsResponse;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestBody;
+
+import java.security.Principal;
+import java.util.List;
 
 public interface MemberControllerDocs {
 
@@ -43,4 +48,33 @@ public interface MemberControllerDocs {
                             content = @Content(schema = @Schema(implementation = ErrorResponse.class)))}
     )
     ResponseEntity<BaseResponse> signUp(@RequestBody MemberProfileCreateRequest request);
+
+    @Operation(
+            summary = "타임 블록 생성",
+            description = "타임 블록을 생성한다.",
+            responses = {
+                    @ApiResponse(
+                            responseCode = "201",
+                            description = "성공",
+                            content = @Content(schema = @Schema(implementation = SuccessResponse.class))),
+                    @ApiResponse(
+                            responseCode = "403",
+                            description = "접근 권한 없음",
+                            content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+                    @ApiResponse(
+                            responseCode = "404",
+                            description = "유효하지 않은 회원",
+                            content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+                    @ApiResponse(
+                            responseCode = "4xx",
+                            description = "클라이언트(요청) 오류",
+                            content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+                    @ApiResponse(
+                            responseCode = "500",
+                            description = "서버 내부 오류",
+                            content = @Content(schema = @Schema(implementation = ErrorResponse.class)))}
+    )
+    ResponseEntity<SuccessResponse<List<BelongTeamsResponse>>> showBelongTeam(
+            @Parameter(hidden = true) Principal principal
+    );
 }
